fix(steadyhand): unify error handling and standardize imports
- Add setError + toast.error to both error paths in ExperienceForm
- Extract error utilities to shared lib/errors.ts module
- Use consistent deep path import for Toaster in layout
